Use ResponsiveImageGroup from membership common 213
