+++ frontend/app/controllers/Info.scala
@@ -2,15 +2,17 @@ package controllers
 
 import com.gu.i18n.CountryGroup
 import com.gu.i18n.CountryGroup._
+import com.gu.memsub.images.{Grid, ResponsiveImage, ResponsiveImageGenerator, ResponsiveImageGroup}
 import configuration.CopyConfig
 import controllers.Redirects.redirectToSupporterPage
+import model.{ContentItemOffer, FlashMessage, Nav, OrientatedImages}
 import forms.MemberForm._
-import model._
 import play.api.mvc.Controller
 import services.{AuthenticationService, EmailService, GuardianContentService, TouchpointBackend}
 import views.support.{Asset, PageInfo}
 import com.netaporter.uri.dsl._
 import com.netaporter.uri.Uri
+
 import scala.concurrent.Future
 import utils.RequestCountry._
