+++ frontend/app/model/Video.scala
@@ -2,5 +2,5 @@ package model
 
 case class Video(
   srcUrl: com.netaporter.uri.Uri,
-  posterImage: Option[model.ResponsiveImageGroup]
+  posterImage: Option[com.gu.memsub.images.ResponsiveImageGroup]
 )
