+++ frontend/app/views/fragments/media/slideshowFeatured.scala.html
@@ -1,4 +1,4 @@
-@(images: Seq[model.ResponsiveImageGroup])
+@(images: Seq[com.gu.memsub.images.ResponsiveImageGroup])
 
 <div class="slideshow slideshow--featured js-slideshow" data-slideshow-autoplay="true">
     <div class="slideshow__container">
