+++ frontend/app/views/fragments/media/slideshowMinimal.scala.html
@@ -1,4 +1,4 @@
-@(images: Seq[model.ResponsiveImageGroup], classes: Seq[String] = Seq.empty)
+@(images: Seq[com.gu.memsub.images.ResponsiveImageGroup], classes: Seq[String] = Seq.empty)
 
 <div class="slideshow @classes.mkString(" ") js-slideshow">
     <div class="slideshow__container">
