+++ frontend/app/views/fragments/page/sectionCardStack.scala.html
@@ -1,4 +1,4 @@
-@(img: model.ResponsiveImageGroup, leadIn: Option[String] = None)(content: Html)
+@(img: com.gu.memsub.images.ResponsiveImageGroup, leadIn: Option[String] = None)(content: Html)
 
 <section class="page-slice page-slice--listing page-slice--slim">
     <div class="l-constrained">
