+++ frontend/app/views/fragments/promos/promoPrimary.scala.html
@@ -1,6 +1,6 @@
 @(
     title: String,
-    img: model.ResponsiveImageGroup,
+    img: com.gu.memsub.images.ResponsiveImageGroup,
     showImageCaption: Boolean = false,
     stampImageOpt: Option[String] = None,
     isLead: Boolean = false,
