+++ frontend/app/views/fragments/promos/promoSecondary.scala.html
@@ -1,6 +1,6 @@
 @(
     title: Html,
-    img: model.ResponsiveImageGroup,
+    img: com.gu.memsub.images.ResponsiveImageGroup,
     showImageCaption: Boolean = false,
     stampImageOpt: Option[String] = None,
     toneClass: Option[String] = None
