+++ frontend/app/views/fragments/promos/promoTertiary.scala.html
@@ -1,7 +1,7 @@
 @(
     title: String,
     description: String,
-    images: Seq[model.ResponsiveImageGroup],
+    images: Seq[com.gu.memsub.images.ResponsiveImageGroup],
     linkOpt: Option[views.support.Links.BasicLink]
 )
