+++ frontend/app/views/index.scala.html
@@ -8,7 +8,7 @@
 @(
   images: model.OrientatedImages,
   catalog: MembershipCatalog,
-  pageImages: Seq[model.ResponsiveImageGroup],
+  pageImages: Seq[com.gu.memsub.images.ResponsiveImageGroup],
   eventCollection: model.RichEvent.EventBrandCollection
 )(implicit countryGroup: CountryGroup)
