+++ frontend/app/views/info/supporterUSA.scala.html
@@ -5,7 +5,7 @@
 @import views.support.PageInfo
 @(heroImages: model.OrientatedImages, supporterPlans: PaidMembershipPlans[Current, Supporter],
   pageInfo: PageInfo,
-  pageImages: Seq[model.ResponsiveImageGroup])(implicit token: play.filters.csrf.CSRF.Token, countryGroup: CountryGroup)
+  pageImages: Seq[com.gu.memsub.images.ResponsiveImageGroup])(implicit token: play.filters.csrf.CSRF.Token, countryGroup: CountryGroup)
 
 @import configuration.Videos
 @import views.support.Asset
