+++ frontend/app/views/promotions/singlePricePlanDiscountLandingPage.scala.html
@@ -15,7 +15,7 @@
 @import views.support.PageInfo
 @import model.Highlights
 @import model.PackagePromo.CtaButton
-@import model.ResponsiveImageGroup
+@import com.gu.memsub.images.ResponsiveImageGroup
 @import com.netaporter.uri.dsl._
 @import views.model.Roundel
 @import com.gu.salesforce.PaidTier
