+++ frontend/app/views/promotions/singleTierLandingPage.scala.html
@@ -18,7 +18,7 @@
 @(pricePlans: PaidMembershipPlans[Current, PaidTier],
     pageInfo: PageInfo,
     headerImage: model.OrientatedImages,
-    promotionImage: model.ResponsiveImageGroup,
+    promotionImage: com.gu.memsub.images.ResponsiveImageGroup,
     promoCode: PromoCode,
     promotion: Promotion[PromotionType, Id, MembershipLandingPage]
 )(implicit countryGroup: CountryGroup)
