+++ frontend/app/views/welcome.scala.html
@@ -1,5 +1,5 @@
 @import views.support.PageInfo
-@(pageInfo: PageInfo, pageImages: Seq[model.ResponsiveImageGroup])
+@(pageInfo: PageInfo, pageImages: Seq[com.gu.memsub.images.ResponsiveImageGroup])
 
 @import configuration.Links
 @import views.support.Asset
